x64: idSIMD_Generic::UpSamplePCMTo44kHz: Program received signal SIGSEGV, Segmentation fault

+++ neo/sound/snd_system.cpp
@@ -344,7 +344,7 @@ void idSoundSystemLocal::Init()
 	}
 
 	// make a 16 byte aligned finalMixBuffer
-	finalMixBuffer = (float *)((((int)realAccum) + 15) & ~15);
+	finalMixBuffer = (float *)((((intptr_t)realAccum) + 15) & ~15);
 
 	graph = NULL;
